Adding Unit testing on timesheet service

+++ src/main/java/com/hgutierrezg/training/service/TimesheetService.java
@@ -1,6 +1,7 @@
 package com.hgutierrezg.training.service;
 
 import com.hgutierrezg.training.dto.TimesheetDto;
+import com.hgutierrezg.training.exception.InvalidRequestException;
 import com.hgutierrezg.training.mapper.TimesheetObjectMapper;
 import com.hgutierrezg.training.model.TimesheetEntity;
 import com.hgutierrezg.training.repository.TimesheetRepository;
@@ -8,6 +9,7 @@ import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -21,7 +23,12 @@ public class TimesheetService {
     }
 
     public void updateTimesheet(TimesheetDto timesheetDto) {
-        this.save(timesheetDto);
+       Optional<Long> id = Optional.ofNullable(timesheetDto.getId());
+       if (id.isPresent()) {
+           this.save(timesheetDto);
+       } else {
+           throw new InvalidRequestException("Id is required for update");
+       }
     }
 
     public List<TimesheetDto> getTimesheets() {
@@ -32,7 +39,11 @@ public class TimesheetService {
     }
 
     public void deleteTimesheet(Long id) {
-        timesheetRepository.deleteById(id);
+        if (Optional.ofNullable(id).isPresent()) {
+            timesheetRepository.deleteById(id);
+        } else {
+            throw new InvalidRequestException("Id is required for delete");
+        }
     }
 
     private TimesheetEntity save(TimesheetDto timesheetDto) {
